+++ frontend/layouts/main.php
@@ -6,20 +6,6 @@ if (!isset($props['styles'])) $props['styles'] = [];
 $props['styles'][] = $parent_call;
 $props['styles'][] = 'counters';
 
-/**
- * Get all the sidebar links from the content
- * 
- * @param string $content The content to parse
- * 
- * @return array<string,string[]>
- */
-function get_sidebar_links(string $content): array
-{
-    $sidebar_links = [];
-
-    return $sidebar_links;
-}
-
 ob_start(); ?>
 
 <!-- TODO: Generate ID for all the headers?!, use js for that? -->