+++ CHANGELOG.md
@@ -21,6 +21,7 @@ All the log of changes on the project/repository
 - Extract sitemap `<url>` node generation into a separate function
 - Center main entry title and subtitle in a separate file
 - Visually redesign, and implement a kind of "glassmorphism"
+- Extract `get_sidebar_links` to `library`
 
 ### Fixed
 - `get_public_url` didn't work for `build` or `img` paths